Proper M-regression for GBM according to Friedman paper https://statweb.stanford.edu/~jhf/ftp/trebst.pdf (Algorithm 4).
GBM now has huber_alpha (the per-leaf-node quantile of |actual-predicted| to find the threshold between quadratic and linear loss).

DL now has huber_delta (directly specify the threshold between quadratic and linear loss relative to standardized response).

NOTE: The mean residual deviance on train/test sets (probably best interpreted as the 'mean huber loss') for DL is computed using the user-given huber_delta, but for GBM, the huber_delta is 1 by default for the computation of the deviance, so it might not make sense. Better to use MAE or MSE/RMSE for GBM's model scoring.